refactor: extract shared frontmatter-scalar reader
feature-source.ts's frontmatterScalar and health.ts's coverageTicketScalar
were byte-identical logic introduced independently in the same PR; unify
into utils/frontmatter.ts.
