+++ packages/cli/src/utils/feature-source.ts
@@ -3,29 +3,17 @@ import nodePath from 'node:path';
 
 import type { PhaseAnchorScope } from '../../templates/hooks/lib/phase-provenance.js';
 import { resolveConfiguredLaneDirectory, resolveTicketsDirectory } from './configured-paths.js';
+import { readFrontmatterScalar } from './frontmatter.js';
 import { readFileSafe } from './fs.js';
 import { toRepoDirectory, toRepoPath } from './repo-path.js';
 import { WORKSPACE_ROOTS } from './workspaces.js';
-
-function frontmatterScalar(content: string | undefined, field: string): string | undefined {
-  const lines = content?.split(/\r?\n/) ?? [];
-  if (lines[0] !== '---') return undefined;
-  const prefix = `${field}:`;
-  for (const line of lines.slice(1)) {
-    if (line === '---') return undefined;
-    if (!line.startsWith(prefix)) continue;
-    const value = line.slice(prefix.length).trim();
-    return value === '' ? undefined : value;
-  }
-  return undefined;
-}
 
 /** Read the authoritative ticket slug, falling back for legacy folders. */
 function slugForTicket(cwd: string, ticketFolder: string): string {
   const content = readFileSafe(
     nodePath.join(resolveTicketsDirectory(cwd), ticketFolder, 'ticket.md'),
   );
-  const slug = frontmatterScalar(content, 'slug');
+  const slug = readFrontmatterScalar(content, 'slug');
   if (slug !== undefined && slug !== '') return slug;
 
   const dashIndex = ticketFolder.indexOf('-');
